requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
weaviate-client==3.24.1
openai==1.3.5
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
